emacs: l/buffer-looks-like: fix call to (or)
This is a followup to the solution in
4d462fc4f6a2734c50db6d38dba0d63b2f6946d5.

The issue is that `or' does not take on a list as an arugment. It takes
on a plain sequence of arguments. So, we have to evaluate (reduce the
booleans) to a single boolean value.

The big gotcha is that `(or (nil))` evalues to TRUE because `(nil)` is
considered a list. We were running into this issue for
l/buffer-looks-like whenever we had just a single regular expression to
comb through.

This solution here simply converts

    (or (arg1 arg2 arg3 ...))

into

    (or arg1 arg2 arg3 ...)

. See
https://emacs.stackexchange.com/questions/62727/how-do-i-splat-the-arguments-to-the-or-function
for more discussion.