emacs: l/quit-buffer: make aux-buffer-rgx more inclusive
Some auxiliary buffers are hidden, such as " *Echo Area 0*" and
" *Minibuf-0*", and so we need to expand the regex to account for these
as well. The risk is that otherwise we end up possibly continuously
running (prev-buffer) forever in the while loop.

For good measure, we modify the while loop here so that if we detect a
cycle, we break execution. This takes care of the case where the
auxiliary buffer regex is not inclusive enough in the future, in the
case where there is some other hidden buffer that is listable with
`(buffer-list)' but not reachable via `(previous-buffer)'.

Also, modify some of the messages sent to the echo area.