zsh: prompt: prevent redraw if we type anything into the buffer
The only time this really comes up as an isuse at all is if we type
something in and then want to subsequently copy the typed-in text from
our terminal emulator. Because Zsh doesn't know about the terminal's
selections, it still overwrites the prompt, resulting in the terminal's
selection disappearing.

Disabling the prompt refresh whenever there's text is a simple
workaround.