zsh: f: make searching for '-*' patterns more intuitive
Before we had to do

    $ f t --pattern --

if we wanted to search for a literal '--pattern'. This makes it so that
we can just do

    $ f t --pattern

(or just 'ft --pattern', using the existing 'ft' alias) instead without
the trailing '--'.

Also, any other arguments to rg are specified before the PATTERN, which
feels more natural. E.g.,

    $ ft --some --other --args --to --rg PATTERN

is now valid syntax.