alacritty: use bigger dimensions
The 163 is because we want to split a terminal emacs window vertically
with enough room for the resulting windows. Each window has 81 columns
because it should have 80 for text and 1 column for git-gutter (162
columns). On top of that we need 1 column dedicated for the split
itself, for a total of 163.