alacritty: mac: use Tamsyn8x16 font
This is the only known font (perhaps in the world?) where we can get
crisp, down-to-the-pixel bitmap rendering without any anti-aliasing on
the font for the Mac. If Terminus had a version for the Mac I would
probably use that for consistency, but there is no such version of
Terminus so Tamsyn is the next-closest substitute.