emacsclient: always show current line position
When a buffer is opened for a 2nd time (that is, the buffer is already
open in the emacs server), we have to first move the cursor around
before the current line is highlighted by hl-line. This is probably
because the server doesn't know which client actually has focus, and so
waits for a client to send a keystroke before designating which client
"holds" the buffer and thus deserves to be "active", thereby triggering
hl-line to kick in from its idle state.

Invoking hl-line-highlight-now is a reasonable workaround.