emacs: prefer to use \\' instead of $
The `$' symbol matches the end of a line. This means that it must be
followed by a newline. The special `\'' construct matches the empty
string, but only at the end of the string, which is what we want.

In other words, "\\.foo$" can match a filename that looks like

    "file.foo\nsomething.txt"

(without quotes) which has a newline that is preceded by ".foo". It is
true that newlines are rare, but still, it is incorrect.