xmonad: use proper font name (crash fix)
There's some funny business going on with libXft. Upon a system
upgrade (new nvidia drivers, etc), for some reason even after we
recompile the xmonad binary, we get a hard crash of xmonad if we try to
refer to the previously-working "dejavu sans mono" font. The workaround
is to use the proper name found with `fc-list | grep deja`.

The hard crash is difficult to work with at first because it's not
immediately obvious (at first sight) when we do need to draw text on the
screen, thus requiring the font.

For context, see https://github.com/xmonad/xmonad/issues/194.