nixos: m0: add to existing k8s cluster on k0
Importantly, we have to open up 2 ports in the firewall for the k8s
master (and reapply on the master with "sudo nixos-rebuild switch").
This is required because we need the k8s node to talk to the master.

The first port is "6443" which is the master API server port. The second
is the "8888" port which is used as the default port for the "cfssl"
service that runs on the master.