tmux: create bindings for next and previous tmux pane
For reasons unknown, invoking "C-M-u" with "xxd -psd" resulted in no
output. So we instead tried "C-M-e", which gave an output sequence of
just "1b05", which corresponds to ESC followed by ENQ (0x05).
Incidentally, the ENQ key is on the same line as the capital "E"
key (0x45) if we look at ascii(7):

       Oct   Dec   Hex   Char                        Oct   Dec   Hex   Char
       ────────────────────────────────────────────────────────────────────────
       000   0     00    NUL '\0' (null character)   100   64    40    @
       001   1     01    SOH (start of heading)      101   65    41    A
       002   2     02    STX (start of text)         102   66    42    B
       003   3     03    ETX (end of text)           103   67    43    C
       004   4     04    EOT (end of transmission)   104   68    44    D
       005   5     05    ENQ (enquiry)               105   69    45    E    <--------
       ...

By imitating the same pattern, for "C-M-u" we deduce that the NAK
key (0x15) is the key that we want, as per this line in the table:

       Oct   Dec   Hex   Char                        Oct   Dec   Hex   Char
       ────────────────────────────────────────────────────────────────────────
       ...
       025   21    15    NAK (negative ack.)         125   85    55    U