tmux: use prefixed keys for pane navigation
The trouble with fc4de04da78fa9f93d79a85b73fdbe1101bfd71f is that the
keys are not prefixed and are native keys. This prevents us from using
these keybindings in a composable way when we are dealing with nested
tmuxes (e.g., when one of our panes is itself a tmux session). Using the
standard prefix key table instead of the root table allows us to use the
existing pattern of invoking C-b multiple times to get into the inner
tmux process.