tmux: show "#{window_flags}" in window list
These window flags include things like a "Z" character when a pane is
zoomed, which is useful to know (as there is no other way to determine
visually if there is a zoomed paned among multiple panes, or if there is
only 1 non-zoomed pane).

With this change we have to adjust the comments because we do again see
asterisks and dashes (again).