tmux: be more flexible about rebalancing panes
Apparently, selecting `even-vertical' as the layout forces all panes to
be vertically stacked on top of each other like pancakes, if you attempt
to create a window "horizontally" with "C-b %".

The new hooks make it behave more like a tiled layout. It is superior to
selecting the `tiled' layout directly (with "select-layout tiled")
because that results in all panes becoming the same width and height as
much as possible. Instead, just doing "select-layout -E" as in this
change makes nearby neighboring panes resize evenly without affecting
panes that are "far away".