tmux_mouse_word_handler.sh: understand specially-named directories
This allows us to easily click on mouse words that are absolute
filesystem paths, including named directories as well. These mouse words
come up all the time because we print the absolute path as part of our
Zsh prompt.

The nice thing is that we send in the TAB key so that we can complete
paths that look like ~/s/foo/bar to resolve to ~/syscfg/foo/bar.