tmux_mouse_word_handler.sh: refactor
For some reason different parts of this code fire. It seems like there's
a race between what "#{mouse_word}" could be at any given moment. We're
still trying to find out what is the cause behind the non-determinism,
but for now this change at least makes the code easier to read.