tmux_mouse_word_handler.sh: fix caching bug
The tmux manpage has this to say about the "#(...)" format specifier:

    When constructing formats, tmux does not wait for ‘#()’ commands to fin‐
    ish; instead, the previous result from running the same command is used,
    or a placeholder if the command has not been run before.  If the command
    hasn't exited, the most recent line of output will be used, but the status
    line will not be updated more than once a second.

So relying on "#(...)" to figure out what to execute based on the
\#{mouse_word} and #{pane_current_path} resulted in cached commands. The
caching resulted in buggy behavior.

This change removes the reliance on "#(...)" and instead simply executes
the tmux commands directly in this script (instead of generating code to
be interpreted by "#(...)"). This is better because it is also simpler
to read (less quoting) and also makes the use of shell variables more
consistent in the script (instead of using a combination of shell
variables and tmux variables of the form "#{...}").