zsh: use zpm, not zinit
Zinit had a strange bug with `@zinit-scheduler()', where on Mac it would
result in "zsh: illegal hardware instruction" if we ran zsh within zsh.
It was probably the case that this was also happening at the outer zsh,
but it would crash so the output would get lost. The workaround on Mac
was to simply define `@zinit-scheduler()' to be an empty function.
However this resulted in plugins not being loaded at all.

An attempt to debug the upstream function was abandoned because the code
is unreadable. This made me lose confidence in the code health of the
rest of the zinit project, and so I opted for "zpm" instead. It has some
quirks like needing to run "zpm clean" every time we update ~/.zshrc,
but this is a small cost to pay.