emacs: add doom emacs as a separate standalone configuration
We use git submodules to pull in chemacs2 (see
https://github.com/hlissner/doom-emacs/blame/bf8495b4122701fb30cb6cea37281dc8f3bedcd0/docs/getting_started.org#L562),
which allows us to have multiple full-blown emacs configurations.

We use this functionality to install the doom emacs distribution (also a
git submodule), because trying out new things is fun and educational.

The end goal is to migrate most of the old functionality over to the
doom configuration and deprecate the "legacy" emacs folder. The
migration presents gives us an opportunity to try to create a literate
configuration file setup.