emacs: add "CSI u" (libtickit) support
Finally we can bind C-S-h and C-S-l in terminal emacs and have them be
differentiated against C-h and C-l. The trick is to (1) make Alacritty
send `CSI u` codes, (2) prevent tmux from sending the non-`CSI u` codes
for C-S- bindings, and (3) extend `CSI u` support from emacs.

Terminal emacs now works fine with these bindings when run from either
inside tmux or without (from Zsh directly). Our terminal Emacs can now
distinguish between things like C-i and C-Tab. See
https://gist.github.com/gnachman/b4fb1e643e7e82a546bc9f86f30360e4 and
http://www.leonerd.org.uk/hacks/fixterms/ for reference.

On the other hand, because Zsh does not support `CSI u` bindings, we
have to basically always call it from inside tmux for the plain C-
bindings to work (such as C-c and C-r). We are able to hack in C-r
support into raw Zsh from Alacritty, but that's about it because of
limitations of Zsh's `bindkey` builtin. But this is a small price to pay
as we hardly ever use Zsh without tmux. For those systems where we just
prefer to use Alacritty+Zsh without tmux, we'll have to update those
hotkeys to always spawn tmux in the middle to avoid the raw Zsh pitfall
described above; this is a small price to pay for a ton of terminal
emacs convenience.