alacritty: avoid `CSI u` bindings for legacy applications
We only use `CSI u` mode for bindings (e.g., C-S-<key> and C-M-S-<key>)
that are not supported in legacy applications that don't support `CSI u`
mode.

A side benefit of this is that we can press C-c in either tmux or in
plain Alacritty, and they both do the same thing. That is, we don't need
the special C-r binding for Zsh any more.

Functionally this is superior because we don't have to special-case
things any more (such as in Zsh).