emacs and tmux: clarify comments
Emacs can indeed see the keys that we previously thought were not being
seen. The issue was that Emacs was just defaulting to the nearest
similar binding when these bindings were not explicitly bound. This
might be a bug, but as we now know that the combinations like M-S-DEL
and others are indeed recognizable by Emacs, we remove these misleading
comments. The lesson here is that we shouldn't rely on `describe-key' to
determine if Emacs can recognize a key combination; rather we have to
bind it first to a command and __then__ run `describe-key' against it.

Also, encode M-Tab with `CSI u` mode for symmetry with the other special
keys like Backspace and Enter.