mac: add karabiner configuration
We use the useful "capslock as hyper key" combination, where the hyper
key is defined as "command+control+option+shift".

From the Mission control settings, we then change the default control
key in the C-right, C-left, C-up, and C-down bindings to this hyper key
and hjkl keys, so that we don't have to leave the home row for switching
between workspaces.

Also, if we just press the hyper key alone with no other letter
combination, it is treated as pressing Escape.