alacritty: have separate keybindings for tmux vs shell
For now the two are identical, but this allows us to use separate
keybindings for alacritty instances that use "exec tmux ..." vs those
that just use "zsh". In this manner we can work around the bug in
https://github.com/tmux/tmux/issues/2705 by making alacritty send a
special key sequence to tmux, and then have tmux interpret that
sequence, as per
https://stackoverflow.com/questions/25294283/bindings-with-key-sequences.
For example, we can make C-i from alacritty/keybindings_tmux.yml send
the sequence "S-F12, i, F12", then make tmux interpret S-F12 as "enter
workaround mode" key table, then make "i" in this key table send the
correct `CSI u` mode "C-i" binding, then make "F12" from this key table
exit this key table.

It may be that we have to use different keybindings than the sequence
described above but that is the general idea. This way we can work
around any such regressions in tmux in the future, which makes it more
robust.

The only technical debt behind this is that if we run tmux manually from
an Alacritty instance launched with alacritty/keybindings_shell.yml,
then the scheme described above won't work. But this is a minor
annoyance.