emacs: terminal: disambiguate between C-i and TAB
We also disambiguate between the following pairs:

- C-m vs Enter
- C-[ vs Escape

This works even from terminal Emacs! The trick is to use a special
Alacritty (terminal) session that sends unique keybindings through to
tmux, and finally to Emacs as well. Each stage of transition (Alacritty
-> tmux -> emacs) requires some careful tweaks. But for all intents and
purposes this is probably as good as it gets for terminal keybindings.

The only caveat is that if you spam these bindings, there's a chance
that either tmux or emacs will interpret the key sequences incorrectly.
However this requires heavy spamming of these keys and generally
speaking this gotcha is virtually harmless.