tmux: make M-ESC send `CSI u` encoded sequence
We typically use M-ESC only from (doom) Emacs, to invoke
`keyboard-escape-quit`, which is by default bound to `M-ESC ESC`. For
whatever reason we could not get it to work with the old binding. With
this new binding, we hold Meta (Alt) and then tap ESC 2x to invoke
`M-ESC ESC`.