zsh: w0: set TERM to xterm-24bit
For whatever reason, using alacritty-xtermlike does not work even if we
install the terminfo entries. This is a semi-broken workaround for
now (24-bit colors don't actually work).

If we don't do this, we run into an odd issue where the panes get
littered with text like

    =2s

which appears to be a tmux thing (it is observable in tmux server logs).
This issue might be relevant: https://github.com/wez/wezterm/issues/999.