emacs: after splitting, change to new window
When we split vertically so that there's a new window below our current
window, in practically every situation we want to move to it to perform
edits because:

      - we have a habit of looking at reference material at the top;
      - our monitors' portrait orientation makes it easier to look at
        things near the middle/bottom of the screen rather than the top.

For horizontal splits, we want to use the new window on the right to do
our edits (and use the left side as a reference) because we deal with
left-to-right text all the time and so the left side tends to remain
"stable" visually after the split.

For these reasons, we move to the newly-created window (both vertically
and horizontally) post-split.

This also matches the focus-shifting behavior of our tmux panes when we
split them.