nixos: add w0 configuration
Run "sudo make nixos" and then "sudo nixos-rebuild switch" to get the
configuration working.

NB: It seems to be the case that resolv.conf cannot be a symlink.

Also, it's worth mentioning here that we are unable to send the Shift
modifier key when SSH-ing with Alacritty, which is somewhat unfortunate.
This is probably due to one of the layers of the stack doing a silent
key/modifier translation, as suggested on
https://github.com/microsoft/terminal/issues/86#issuecomment-407276514.
This isn't a total dealbreaker for us (at least not yet).