zsh: rework 't' alias
The 't' alias can be used to invoke the new 't' function (overwriting
previous behavior of calling syscfg/script/tmux.sh). This function
creates a new tmux session with the lowest free session id, and then
switches the current client over to it. This is useful if we connect to
a remote session over SSH (which already has a tmux session), and if we
want to create a separate, new tmux session on the remote (because with
the current setup, SSH-ing into the remote with a new window will simply
reattach to an existing tmux session).

Also move all custom Zsh functions into lib.sh for better organization.