redo terminal colors
Make all themes use foreground and background colors from the 16
colors (namely black, white, brightblack, brightwhite). This and some
other changes in tmux make it so that changing colors (regardless of
whether we're in tmux or not) do not result in white text on a white
background, for example. The colors are now consistent when we use the
'c' and 'C' color changing aliases in Zsh.

We also use a new Dhall script, renderAlacrittyColorYaml.dhall, to
generate colors for Alacritty. This makes things more uniform (and opens
up the possibility of changing colors for new Alacritty instances), as
we now just use a symlink to choose the default startup color.