tmux: use terminal theme colors
We also reuse cursor and text colors for the tmux status bar, because
this gives us 2 extra colors that don't collide as much against the
black/brightblack colors defined in the 16-color palette of the theme.

Technically speaking we can define unlimited colors because the tooling
stack we use supports 24-bit colors. That is, we can define an arbitrary
number of 24-bit colors and just refer to them using the same `tmux
set-environment ...` scheme.