tmux: introduce passthrough mode
Pressing C-M-b enters `passthrough` mode (i.e., the `passthrough`
key-table). This mode is designed to ignore all keypresses and instead
pass them through to the inner (nested) tmux instance. Even the C-b
prefix key is disabled so that it gets passed through to the inner tmux
instance.

As for implementation, the main thing is that we have to re-encode
C-S-... and C-M-S-... keys, but other than that it works without
fanfare. We press C-M-b again to exit this mode to regain control of the
outer tmux session.