emacs: replace all hardcoded colors with dynamic colors
We make the zsh `set_theme` function invoke the new `l/reset-faces`
function, so that all faces get regenerated with the new colors from the
theme. As for the initial set of custom colors, we use PastelDark.dhall
because that's the same color scheme used in the default Alacritty
invocation.