nixos: merge in nixpkgs configuration
Most of these packages are only updated maybe once or twice a year, so
it makes sense to just define it at the system level. If we need to use
a newer version of a package, we can install it at the user level and
then wait for the nixos channel to catch up.