lh: tick: do not delete repo_stats outright
Instead use a boolean to mark it as stale. This avoids cases where the
cached repo_stats is invalidated and we serve a "{}" to lhc (typically
happens when we create or delete a file very quickly, such that the tick
has not had time to re-evaluate the new git repo stats yet), making the
prompt fall back to the gsts() function.