lh: when marking as stale, change status to LOADING
This covers the case when the watcher process is already running, but
has been marked stale. Until the stats have been recomputed, we set its
status to "LOADING" so that any other clients asking for the status know
that it is currently being recomputed.

This is noticeable when we touch/rm some files while already in a Git
repo (esp. a large one).