emacs: org: remove ".org" suffix from buffer names
Most of the time we are using Org files for personal note-taking, and
don't care about the suffix. Remove it so that there is less noise in
the tab bar. We still get the full path and filename in the modeline, so
we don't actually lose any real information here.