emacs: adjust colors
For some reason when the frame loses focus, *some* of the faces in the
modeline are treated as inactive and the other as still active. This
ends up hiding some of these faces (one of the most important ones being
the filename of the buffer).

So instead we always make the foreground color be "base0" for both
active and inactive modelines. We adjust the background color
accordingly, and make identical changes to the tab bar for aesthetic
consistency.