add wezterm
The main motivation for this is a bug in Alacritty where we cannot bind
`C-,` (CTRL+COMMA) in Linux (although it works on Mac), as described here:
https://github.com/alacritty/alacritty/issues/6354. The bug is not a bug
according to upstream (working as intended), so this is a dealbreaker
for us because we use the comma as a leader key in Doom Emacs (instead
of the default Space key) and we would like to enter this leader key
from within insert mode with CTRL+COMMA. Emacs org-mode also uses this
binding for some operations, so this also breaks usability of org-mode
on Linux.

Thankfully, configuring wezterm is quite easy and it appears to be even
more extensible (more things that are configurable) than Alacritty. As
wezterm is also cross-platform and appears to be relatively stable, we
can just make wezterm a drop-in replacement for Alacritty. And finally,
yes, wezterm is able to bind CTRL+COMMA without problems so that's the
proverbial nail in the coffin.

We will delete Alacritty configurations in a later commit.