zsh: tmux: prefer unattached sessions
Previously when opening a terminal window, by default we always created
a new session by finding the lowest unclaimed session number. This was
fine but for cases where we closed the terminal window, the session that
this terminal was looking at was "abandoned" because subsequently
opening a new terminal would just create a new session instead of
opening the unattached session.

This change prioritizes unattached sessions to connect to them first, if
any. IOW, we only create a new tmux session if we detect that all
sessions are attached.

Note that "attached" means either attached with a local terminal window
or a remote SSH session. Treating SSH connections as "attached" in this
sense is fine because we'll just create a new session in the worst case.
This new behavior means that by default we will prefer to not attach
locally to a tmux session that is attached from a remote tmux client
connection.