mool: configure for nix-darwin
We also avoid symlinking because the absolute path is system-dependent.
Instead we just copy in the sample config files that ship with mools
from wherever mools is installed.

We also use slightly broader .gitignore rules to ignore basically
everything inside the mool folder, because we don't want to check these
default files in. Later on if we deviate from the sample config we can
check those in at that time.