zsh: simplify "kl" function
"kl" is no longer a standalone function. This is because we never needed
it to be a function (meant to be called and reused elsewhere). Instead
we just replace the leading word "kl" to "kubectl" immediately. This is
better because now our logs become more explicit.

We should probably do the same for all of the other commonly-aliased
commands, assuming that the growth of the "case" statement used here
remains performant. Explicit is better than implicit.