emacs: disable auto-dim-other-buffers
Sadly it does not play well with the latest doomemacs (if multiple
windows are viewing the same buffer, the backround becomes very
erratic). This probably has something to do with the timers that
auto-dim-other-buffers uses, such as
`adob--focus-change-debounce-delay`. However, it's probably doom's fault
because the code for disable auto-dim-other-buffers hasn't changed in
almost a year.

Hopefully this package can become usable again soon.