mac: disable pressing and holding a key to get another key
At one point (last week? two weeks ago?) pressing and holding the "l"
key would result in typing in an accented "ł" key. This was very painful
and I recall finding the trick in this patch somewhere on the internet.
However it is unclear if this actually works (if it actually fixed the
problem). Still, it's better to include this here at least for
historical context.