emacs: disable doom ui colors
We get behavior similar to what is described in
https://stackoverflow.com/questions/72803319/the-display-is-wrong-when-using-iterm2-screen-emacs-on-macos-monterey,
although the problems show up in just Linux. Basically the display gets
corrupted when we have two different buffers displayed in the same
window, side-by-side, and when we start scrolling around in the buffer
on the right.

It appears that the problem is significantly reduced if we just don't
use the doom-zenburn (or any of the doom color themes). It could be that
the 24-bit colors defined by doom are problematic for the wezterm+tmux
combo. Or it could be that the actual doom color themes are to blame.
It's probably the former, but it's difficult to test these two scenarios
separately. So for now just use the default color theme instead of
doom-zenburn.