emacs: git-gutter: disable refresh-on-frame-focus
Originally, we wanted to make git-gutter refresh when regaining frame
focus because it could be the case that the buffer could be updated
outside of emacs (perhaps when using git directly, or when another
editor like nvim modifies the same file). But the cost is that the
terminal can get messed up pretty badly, requiring a redisplay in order
to clean up the characters that can get garbled.

Instead, just use the built-in timer and make it refresh automatically
based on a time interval.