emacs org-mode: turn off org-indent-mode
We've been using this mode for years, but let's see how it feels without
it. The main benefit is that the overall line lengths are more
consistent because there are no more virtual indents. This makes it more
possible to do vertical splits without overflowing lines on portrait
mode monitors.