zsh: lazy-load "kubectl" completions
Before, the completions only worked for the "kl" shortcut. This was a
problem because when we went back into our shell history, the entries
would have "kubectl" in them (because we modified the "kl" to
"kubectl"), and we wouldn't load the completions.

Load kubectl completions for both "kl" and "kubectl".