tig: use "h" and "l" for parent/child blame navigation
We use blame a lot, and these "parent" and "back" (where we were before
we went to the parent commit, or, the previous child that we were on)
commands are very useful.

The "(" and ")" bindings were originally introduced in d4ffbd08 (tig:
make bindings resemble notmuch-emacs bindings, 2017-11-08).  But they
weren't really used in practice, so remove them.

The C-h and "o" bindings are here to account for the fact that the old
"h" and "l" bindings (for "view-help" and "view-log") were overridden by
the new bindings in this change. Sadly, tig gave us a warning when we
tried to use C-l for "view-log" because it conflicted with the binding
for "screen-redraw" like this:

    tig warning: ~/.tigrc:25: Key binding for screen-redraw and none conflict; keys using Ctrl are case insensitive
    tig warning: Errors while loading /home/l/.tigrc.

so we just used "o" instead for "view-log".