emacs: disable tty module
It currently has a bug on Mac where copying a line with "yy" and then
pasting with "p" (evil-paste-after) results in the contents being pasted
onto the current line (not the next line below the current line).
Disabling this appears to disable "xclip-mode", which is the problematic
mode for us.

The symptoms described above appear to be similar to the ones described
at [1], modulo the lack of clipetty mode (the tty module enables
clipetty mode only if "+osc" argument is set, which we don't do). Still,
at the comment at [2] the solution is to disable xclip-mode. Disabling
the "tty" module appears to disable xclip-mode permanently, which is
what we want.

According to [3], the tty module enables three features:

    * System clipboard integration (through an external clipboard
      program or OSC-52 escape codes in supported terminals).

    * Cursor-shape changing across evil states (requires a terminal that
      supports it).

    * Mouse support in the terminal.

Even without the tty module, system clipboard integration appears to work,
probably because we've already configured Emacs to defer to tmux for
copying into the system clipboard (and we always run Emacs under tmux).
And we don't care about cursor shape changing across Evil states (in fact we
explicitly don't want it) so that's a plus. Lastly we really don't care
about mouse support inside Emacs anyway because we never need it. So it
looks like the tty module was essentially useless for us anyway. So
disable it!

[1] https://github.com/spudlyo/clipetty/issues/8#issue-548507228
[2] https://github.com/spudlyo/clipetty/issues/8#issuecomment-609495496
[3] https://docs.doomemacs.org/v21.12/modules/os/tty/