lieer: add script to fetch emails
Note that this does not cover _sending_ emails (from Emacs). That will
be in the next commit.

Also, we have to tweak the PATH because we will call this from cron with
a rule like

    */3 * * * * /Users/l/syscfg/script/mail-sync.sh

but the cron environment does not read in paths used by the Nix package
manager. So the script has to adjust the PATH to find "lieer" and
"notmuch" on the system.

My memory is hazy but this is what I did to initially set up notmuch and
lieer:

    # Install packages. These will end up in ${HOME}/.nix-profile/bin.
    # If we install them into the system (by specifying them in
    # configuration.nix), they will end up in
    # /run/current-system/sw/bin.
    nix-env -iA nixpkgs.lieer
    nix-env -iA nixpkgs.notmuch

    # Set up notmuch. Choose "~/mail" (default) folder for mail.
    notmuch setup
    mkdir -p ~/mail/linusarver@gmail.com

    # Set up gmail credentials (this will direct your web browser to the
    # OAUTH consent page).
    cd ~/mail/linusarver@gmail.com
    gmi init linusarver@gmail.com

    # Initial sync. From now on you can run "gmi sync" from here.
    gmi sync