notmuch: tag emails in batch mode
Gmail's filters do the job of labeling emails (e.g., moving them out of
the Inbox in the web view if we have "Skip the Inbox" checked for a
filter). lieer will fetch new mails from Gmail and translate certain
labels to tags automatically for us (e.g., if an "unread" tag is
missing, then lieer will mark the same email in Gmail as having been
read in the next sync).

From notmuch, we can do some additional tag processing to make notmuch
searches a bit more convenient. For example, we use a "list" tag that
does not exist as a label in Gmail to search across all emails across
all mailing lists.