nixos: disable RAID-1 drive
This broke during the 22.05 -> 23.11 transition, somehow. Disable it for
now while we figure things out (the drive is mostly empty so whatever).
See [1] for a similar story of breakage. Unfortunately for us we have an
extra hurdle because the drive is encrypted and we have to decrypt it
during stage 1.

[1] https://github.com/NixOS/nixpkgs/issues/272393