tmux: do not hardcode xsel
We already tell tmux to overwrite the clipboard with

    set -g set-clipboard on

so it doesn't make sense to do it again with "xsel" (and of course xsel
is only available in the X window system and is not portable).

Also avoid exiting copy mode after copying, because oftentimes we scroll
up a fair way up the buffer before copying and we want to be
conservative and preserve the context (position in the buffer) so that
we don't have to scroll up again if we want to copy something else close
to it.