emacs orgmode: tweak markdown exporter
The

    (org-export-with-smart-quotes nil)
    (org-export-with-special-strings nil)
    (org-export-with-fixed-width t)

bits make the export a bit more source-control friendly by disabling
special "smart" characters which end up as HTML entities (gross). The
"org-export-with-fixed-width" option is a surprise as that appears to be
the only way where source code blocks are preserved.

The

    (visible-only nil)

allows us to export all headings, not just those that have been
expanded (if they happen to be folded).

Remove

     (flush-lines "^$")

because we want to keep blank lines (we did this originally becuse the
system where we were pushing the contents did not like blan lines).
Keeping the blank lines helps separate paragraphs from each other.

Lastly we use "evil-yank" instead of "~/syscfg/script/copy-clipboard.sh"
because we run Emacs inside of tmux and it already integrates with
the system clipboard.