emacs modeline: truncate more aggressively
When splitting vertically on a Mac display, if we have really long
file paths most of the modeline will display just the file path and
nothing else. It is surprising that doom-modeline doesn't automatically
detect whether some items in the modeline are not visible because of a
narrow window width (and adjust things accordingly to display as much
info as possible).

Set this window width value to a higher number so that anything less
than a width of 100 triggers the "truncate file path" logic to show the
other important things like line and column number. Because we don't
display line/column numbers anywhere else (for simplicity), setting this
value is important.

Historically we've never noticed the need to set this value because we
did not work in Java-like codebases before with super long file paths.