nix-darwin: enable Lorri daemon
Upstream has fixed the issue in version 1.7.0 [1]. Apparently we've
always enabled the (broken) daemon on Linux systems since
e40aa1ca (nixos: add direnv and lorri, 2022-08-08), but never bothered
to add it to Mac with nix-darwin, probably because of the known
breakage.

Add it now because it should be working now. [2] shows
that we can enable the "services.lorri.logFile" option, which allows us
to specify a particular path to use for the log file. We don't set this
because we don't know if this means unbounded growth for that log file
on Mac systems. Interestingly, [2] also mentions an alternate way to to
enable/disable logging the service under launchctl without this option::

   services.lorri.logFile

   The logfile to use for the lorri service. Alternatively "sudo launchctl
   debug system/org.nixos.lorri --stderr" can be used to stream the logs to
   a shell after restarting the service with "sudo launchctl kickstart -k
   system/org.nixos.lorri".

[1] https://github.com/nix-community/lorri/issues/90#issuecomment-2240245933
[2] https://daiderd.com/nix-darwin/manual/index.html