emacs: fix custom dictionary file format
Although we could do "zg" (+spell/add-word) to add to
"custom-dict.txt", when we ran "z=" (+spell/correct), it complained with

    ispell-init-process: Error: The file "/Users/l/syscfg/emacs/spell-fu/custom-dict.txt" is not in the proper format.

According to [1], the solution is to manually add a magic string at the
top of the file. Indeed, this works to suppress the error message.
Although there is official documentation describing this file format
[2], there doesn't appear to be any instructions about automatically
generating this format (perhaps because it's so simple).

[1] https://github.com/doomemacs/doomemacs/issues/6246#issuecomment-1122245805
[2] https://github.com/GNUAspell/aspell/blob/ee6cbb12ff36a1e6618d7388a78dd4e0a2b44041/manual/aspell.texi#L2068-L2082