emacs: disable spellchecking for tables in Org mode
When we run SQL queries inside source code blocks, we don't want to
spellcheck the results (for such things like database column names and
values inside rows).

Interestingly, the `alist-get` function can be used as a "PLACE"
according to Emacs docs. Because `cl-pushnew` can take a "PLACE", we can
use `alist-get` in there directly (as opposed to, say, creating and
returning an augmented, (new) association list). This feels a bit
gross (it's imperative, not functional), but it is what it is.