org-roam: add new folder for storing "join" notes
We use the term "join" like how it means in relational databases. The
notes in zk-join will contain zettels that reference other zettels. They
are also known as "master" or "index" cards in other Zettelkasten
setups, where the point is to list other related zettels (typically in
one category, or "box") in one place.

We prefer to use "join" because it could very well be the case that two
different join cards have significant overlap with each other, such that
neither one is more "authoritative" than the other.