git: simplify force pushes
This setting is a result of Git complaining that it will only push up
branches to a remote if there's a corresponding upstream branch for it
set up on the remote side. Example:

    fatal: The current branch foo has no upstream branch.
    To push the current branch and set the remote as upstream, use

        git push --set-upstream u foo

    To have this happen automatically for branches without a tracking
    upstream, see 'push.autoSetupRemote' in 'git help config'.

This can be a bit annoying if we're in the habit of creating branches
locally first, for review by others on a remote. Set "autoSetupRemote"
to true to silence the warning.